Moved components to be loaded async to the needed file
