+++ src/index.tsx
@@ -10,32 +10,6 @@ import { Router,  hashHistory  } from 'react-router'
 
 console.log("Hello from root index");
 
-// If it needs async loading it will just call require.ensure (webpack)
-const componentStudents = () : any => {
-  // https://medium.com/@Nadav.Dav/implementing-lazy-loading-with-webpack-and-react-router-f8497f895892#.mq4ug0c15
-  return {
-      getComponent: (location, cb) => {
-        require.ensure(['./components/pages/students'], require => {
-              // Retrieve product page component
-              cb(null, require('./components/pages/students')["default"]);
-        });
-      }
-   };
-};
-
-const componentTrainers = () : any => {
-  // https://medium.com/@Nadav.Dav/implementing-lazy-loading-with-webpack-and-react-router-f8497f895892#.mq4ug0c15
-  return {
-      getComponent: (location, cb) => {
-        require.ensure(['./components/pages/trainers'], require => {
-              // Retrieve product page component
-              cb(null, require('./components/pages/trainers')["default"]);
-        });
-      }
-   };
-};
-
-
 // The ...component, spread operator: like object assign just add the new
 // properties to the Route control (and preserves the existing ones)
 ReactDOM.render(
